+++ python_codes/fieldstone_35/fieldstone.py
@@ -135,7 +135,7 @@ for j in range(0,nnr):
 
 print("connectivity")
 
-icon =np.zeros((m, nel),dtype=np.int16)
+icon =np.zeros((m, nel),dtype=np.int32)
 
 counter = 0
 for j in range(0, nelr):
